refactor: create app test file
